Standardize plugin names and APIs
