+++ src/hooks/useEditorEventListener.ts
@@ -3,7 +3,7 @@ import type { DOMEventMap, EditorView } from "prosemirror-view";
 import { useCallback, useContext, useRef } from "react";
 
 import { EditorContext } from "../contexts/EditorContext.js";
-import type { EventHandler } from "../plugins/componentEventListenersPlugin.js";
+import type { EventHandler } from "../plugins/componentEventListeners.js";
 
 import { useEditorEffect } from "./useEditorEffect.js";
